+++ src/AssociationRegistry/EventStore/EventStore.cs
@@ -1,6 +1,5 @@
 namespace AssociationRegistry.EventStore;
 
-using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
 using Be.Vlaanderen.Basisregisters.AggregateSource;
@@ -8,6 +7,7 @@ using Framework;
 using JasperFx.Core.Reflection;
 using Marten;
 using Marten.Exceptions;
+using NodaTime.Text;
 using IEvent = Framework.IEvent;
 
 public class EventStore : IEventStore
@@ -24,7 +24,7 @@ public class EventStore : IEventStore
         await using var session = _documentStore.OpenSession();
 
         session.SetHeader(MetadataHeaderNames.Initiator, metadata.Initiator);
-        session.SetHeader(MetadataHeaderNames.Tijdstip, metadata.Tijdstip);
+        session.SetHeader(MetadataHeaderNames.Tijdstip, InstantPattern.General.Format(metadata.Tijdstip));
         try
         {
             var streamAction = metadata.ExpectedVersion.HasValue ? session.Events.Append(aggregateId, metadata.ExpectedVersion.Value + 1, events.As<object[]>()) : session.Events.Append(aggregateId, events.As<object[]>());